+++ newsroom-lib/src/main/java/com/theguardian/newsroom/desks/DatabaseDesk.kt
@@ -1,9 +1,9 @@
-package com.theguardian.newsroom.eventhandling
+package com.theguardian.newsroom.desks
 
 import com.theguardian.newsroom.archive.EventWriteRepository
 import com.theguardian.newsroom.model.Event
 
-class DatabaseEventHandlingDelegate(private val eventWriteRepository: EventWriteRepository) : EventHandlingDelegate {
+class DatabaseDesk(private val eventWriteRepository: EventWriteRepository) : Desk {
     override fun handleEvent(event: Event) {
         eventWriteRepository.writeEvent(event)
     }
