+++ newsroom-lib/src/main/java/com/theguardian/newsroom/model/Event.kt
@@ -2,6 +2,6 @@ package com.theguardian.newsroom.model
 
 import java.util.*
 
-data class Event(val source: String, val title: String, val data: Map<String, String?>? = null) {
+data class Event(val source: String, val title: String, val data: Map<String, String?>? = emptyMap()) {
     val date = Date()
 }
