--- newsroom-lib/src/main/java/com/theguardian/newsroom/reporter/ReporterTasks.kt
@@ -1,9 +1,0 @@
-package com.theguardian.newsroom.reporter
-
-import com.theguardian.newsroom.Newsroom
-
-interface ReporterTasks {
-    fun setNewsroom(newsroom: Newsroom)
-    fun onStart()
-    fun onStop()
-}
