+++ newsroom-lib/src/main/java/com/theguardian/newsroom/ui/MainActivity.kt
@@ -5,7 +5,7 @@ import android.os.Bundle
 import android.support.v7.widget.LinearLayoutManager
 import com.theguardian.newsroom.Newsroom
 import com.theguardian.newsroom.R
-import com.theguardian.newsroom.reporter.GoogleAnalyticsReporter
+import com.theguardian.newsroom.reporter.GoogleAnalyticsBaseReporter
 import kotlinx.android.synthetic.main.activity_main.*
 
 class MainActivity : Activity() {
@@ -15,7 +15,7 @@ class MainActivity : Activity() {
         setContentView(R.layout.activity_main)
         initRecyclerView()
 
-        Newsroom(this).addReporter(GoogleAnalyticsReporter())
+        Newsroom(this).addReporter(GoogleAnalyticsBaseReporter())
     }
 
     private fun initRecyclerView() {
